Added new game menu, work on issues #32,#33,#34
